+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.7] - 2025-04-06
+
+### Added
+- Changes in TODO.md
+- Changes in examples/filter_files.py
+- Changes in src/infrash_embedded/filter.py
+
 ## [0.1.6] - 2025-04-06
 
 ### Added